remove duplicates from pricing-table-one
